+++ src/main/java/model/user/data/PlayerNames.java
@@ -1,4 +1,4 @@
-package model;
+package model.user.data;
 
 import exception.PlayerNamesOverlapException;
 import utils.StringUtils;
