+++ src/main/java/model/user/data/PlayersData.java
@@ -1,8 +1,9 @@
-package model;
+package model.user.data;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Set;
+import model.user.money.BettingMoney;
 
 public class PlayersData {
     private final Map<String, BettingMoney> playersData = new LinkedHashMap<>();
