+++ src/main/java/model/user/money/BettingMoney.java
@@ -1,4 +1,4 @@
-package model;
+package model.user.money;
 
 import exception.IllegalBettingMoneyFormatException;
