+++ src/test/java/model/user/money/BettingMoneyTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.user.money;
 
 import exception.IllegalBettingMoneyFormatException;
 import org.junit.jupiter.api.DisplayName;
